move the NamedSignalValue class into its own file
