+++ cantools/database/can/formats/sym.py
@@ -24,11 +24,12 @@ from textparser import (
 )
 
 from ...errors import ParseError
+from ...namedsignalvalue import NamedSignalValue
 from ...utils import SORT_SIGNALS_DEFAULT, sort_signals_by_start_bit, type_sort_signals
 from ..internal_database import InternalDatabase
 from ..message import Message
 from ..signal import Decimal as SignalDecimal
-from ..signal import NamedSignalValue, Signal
+from ..signal import Signal
 from .utils import num
 
 LOGGER = logging.getLogger(__name__)
